:truck: (interface): Rename file interface/CoreSDRAM.hpp to interface/SDRAM.hpp
